🚚 api: move serializer to src/certification/configuration context
