+++ api/src/certification/configuration/infrastructure/serializers/complementary-certification-serializer.js
@@ -2,7 +2,7 @@ import jsonapiSerializer from 'jsonapi-serializer';
 
 const { Serializer } = jsonapiSerializer;
 
-import { ComplementaryCertification } from '../../../domain/models/index.js';
+import { ComplementaryCertification } from '../../../../../lib/domain/models/index.js';
 
 const serialize = function (complementaryCertifications) {
   return new Serializer('complementary-certification', {
