+++ api/tests/certification/configuration/unit/infrastructure/serializers/complementary-certification-serializer_test.js
@@ -1,7 +1,7 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/complementary-certification-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/certification/configuration/infrastructure/serializers/complementary-certification-serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
-describe('Unit | Serializer | JSONAPI | complementary-certification-serializer', function () {
+describe('Certification | Configuration | Unit | Serializer | complementary-certification-serializer', function () {
   describe('#serialize', function () {
     it('should convert a ComplementaryCertification model object into JSON API data', function () {
       // given
